Rename module to indra_world
