+++ indra_world/assembly_dashboard/templates/index.html
