+++ indra_world/resources/assembly_compositional_december2020.json
