+++ indra_world/resources/assembly_compositional_phase3.json
