+++ indra_world/resources/default_pipeline.json
