+++ indra_world/resources/groundings_to_exclude.txt
