+++ indra_world/resources/statement_preparation.json
