+++ indra_world/tests/eidos_compositional.jsonld
